Improvements for TwinCAT3: single ams net id, notifications
Major improvements to work with TwinCAT 3 systems.
They only allo one ams net id <-> IP connection.
The basic idea is, to "patch" the invoke id, put it into a hash table
indicating the client, and distribute it to the client when the PLC answers.
(This works similar to NAT in network routers)

Another improvement is the support for notifications:
They are now handled.
Note that different client may ask for the same notifications.
In this case, we request only one notification from the PLC, and "multiply" the
notification into the different clients.
Note that the notifications from the PLC do not use the original invoke id.

Make it possible to specify a local ams net id by using the --local_ams_net_id
option

Other changes:
  For some devices (bus couplers) add_route() will silently do
  nothing and return OK.
  The caller will call
  info!("TCP: added backroute to forwarder");
  even if no route had been added.
  Solution: Do not print here.

main: Do not use colors when stdout is not a tty
  Redirecting stdout to a file (or using tee) leaves the ANSI color
  inside the file, which makes them too hard to read with an editor.
  Remove the coloring when stdout is not a tty.

Add more get_xxx() functions to extract fields from an AMS message
    Avoid peeking at addresses inside the AMS header(s)
    Write proper functions like get_source_id() and use them.

Change-Id: I2480eda4d3348ad8ee1973d7c9c78e6216cb7f8d