handle get_event no client any more ... better
Keep better track about notification, that are deleted:
avoid a memory leak and call
notif_handle_to_last_notif_stream_map.remove(handle)
when a notification is deleted towards the PLC.

And when we get a notification, which is still in the pipe, as
DELNOTIF has not yet reached the PLC and has been processed,
check if we have any last/latest notif_data for this handle.
And no client which wants a notification is found, then, only then
do the print
"get_event no client any more sample.handle",
since there is something wrong.

Change-Id: Ie4b9b912ef4d78ac8af8e44b3055a8bac82c5c59