forwarder: Fix notif_req_data_to_handle_map.remove
Fix a bug:
When a client had more than one notifications, and disconnects,
we clean up the notifications for that client.
If this was the only client with this notification(s),
a DELNOTIF is send to the PLC.

However, the notif_req_data_to_handle_map was not cleaned up
for all notifications. Only for one of them.

This leads to the problem, that the next client did not receive
any notifications.

No new ADDNOTIF was send to the PLC.
The new client should have gotten noticifications from a stale
handle, that never come.

Solution:
After sending out all DELNOTIF to the PLC, collect all notif_req_data
that are stale and clean up notif_req_data_to_handle_map in a loop.

Change-Id: I41c7c09046b8092432f027c691173d4e36ddddd9