util.rs: Make sure that all AdsMessage are debug-printed
The debug!() macro can be helpful to get a decoded printout
of an ADsMessage.
However, some of them are not printed, as an example a "GetState"
message that had the wrong source_id.

Add a default debug!() to catch these cases, when debug_printed
had not been set to true, caused by a "good" debug!()"